fix windows build; more fixed warnings

+++ src/orm/qtsql_driver.cpp
@@ -144,7 +144,7 @@ QtSqlConnectionBackend::open(SqlDialect *dialect, const SqlSource &source)
         conn_->setHostName(source.host());
     String options;
     Strings keys = source.options();
-    for (int i = 0; i < keys.size(); ++i) {
+    for (size_t i = 0; i < keys.size(); ++i) {
         if (!str_empty(options))
             options += _T(";");
         options += keys[i] + _T("=") + source[keys[i]];
